requirements.txt
exif
python-dateutil
Pillow
